finish decidable equality of type, now do he same for exp
Got sidetracked because I got sucked into writing Wiki pages for Coq...
Link: https://github.com/coq/coq/issues/7788

Was trying to understand Canonical Strcutures v/s Typeclasses.

This came up because I wanted an easy way to show decidable equality,
and it was suggested on the Gitter channel that I establish this by
showing an injection into `mathcomp.choice.gentree` which has many
lemmas proved on it.