Find correct theorem statement needed to split traces into chunks of memory effects.
The idea is to set the type of `mem_effect` to
`mem_effect :: Memory -> Trace X -> Memory * Trace X`

That way, we can "consume" as much of the trace as required, and then
our theorem statement becomes:

memD initmem (x >>= f) ~= let (m', x') = mem_effect initmem x in
    memD m' (x' >>= f)

So, we return _both_ the new state of memory _and_ the new trace.
Note that we currently need a witness that the trace is finite.

I feel there is another formulation of this theorem that could work,
in the presence of infinite traces, by wrapping this into a coinductive?
Something like,

`mem_effect :: Memory -> Trace X -> Trace (Memory * X)` (?)
with the rule:

memD initmem (x >>= f) ~=
(mem_effect x initmem) >>=
    (fun (m', x') => memD m' (x' >>= f))

But anyway, don't need this right now so...