Destroy our memory model to give up on external function calls.
This is useful since we can now talk about enriching traces with the
value of memory. That is, we can write a function:

mem_effect: Mem -> Trace X -> Trace (Memory * X)

where the leaf nodes are enriched with the state of memory at that
point. The downside is that our memory model gives up when there is an
external function call and returns am "Err"

The less destructive path would have been to have a coinductive
predicate (in Prop) `TraceDoesNotContainExternalFunctionCall`
that the trace does not contain external function
calls. So we would write a function:

mem_effect: Mem -> (x: Trace X)
                -> TraceDoesNotContainExternalFunctionCall x
                -> Trace (Memory * X)

However, Tobias had a convincing argument that this extra proof
engineering is not worth the effort, which I agree with. We can amend
this later if need be.

This is nice, because we can now talk about memory effects on infinite
traces without having to provide an explicit witness (in Type) about the
finiteness of the trace.