Show loop execution from backedge.
This is a "large" proof term, because it contains all the different
paths that can happen: depending on the relationship between the
indvar and the upper bound of the loop, the program can do different
things. The theorem I just showed shows "all paths" with conditionals.

We can specialize this theorem to show what happens when we execute this
loop under certain preconditions (eg: for the final loop iteration, for
the intermediate loop iterations...)