Proved: LastWriteImposesMemoryValueAtLastWriteIx_scop_stmt
That is, if we know that there is a last write, and we are reading
from the last write's memory location at the point in time where the
last write happens, then the value you receive will be that of the last
write. Sounds simple, it is, but it needs some auxiliary concepts to
pin down precisely. All in all, 8/10 fun proof.